[prometheus] Support the prometheus openstack exporter
This patch implements the initial support for the
openstack-exporter[0] in the kolla-ansible
prometheus monitoring system.

The configuration and prechecks are reused from the other
exporters and a new template is provided for generating
a os-client-config file required by the exporter.

The default scrape interval is 60 seconds, but it can
be extended via a configuration option.

[0] https://github.com/Linaro/openstack-exporter

Change-Id: I4a34c4bb56e74b5cd544972cbd6540d9acb6e4a1

Conflicts:
	ansible/group_vars/all.yml
	ansible/roles/prometheus/defaults/main.yml
	ansible/roles/prometheus/templates/prometheus.yml.j2
	ansible/site.yml
	etc/kolla/globals.yml